test: move type aliases and classes out of conftest modules
