+++ tests/integration/api/v1/providers/test_providers.py
@@ -6,10 +6,9 @@ import pytest
 from fastapi.testclient import TestClient
 
 from merino.providers import BaseProvider
-from tests.integration.api.v1.conftest import (
-    NonsponsoredProvider,
+from tests.integration.api.v1.models import NonsponsoredProvider, SponsoredProvider
+from tests.integration.api.v1.types import (
     SetupProvidersFixture,
-    SponsoredProvider,
     TeardownProvidersFixture,
 )
