+++ tests/integration/api/v1/suggest/test_wiki_fruit.py
@@ -7,7 +7,7 @@ from fastapi.testclient import TestClient
 
 from merino.providers import BaseProvider
 from merino.providers.wiki_fruit import WikiFruitProvider
-from tests.integration.api.v1.conftest import (
+from tests.integration.api.v1.types import (
     SetupProvidersFixture,
     TeardownProvidersFixture,
 )
