+++ tests/unit/conftest.py
@@ -2,14 +2,11 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-from typing import Callable
-
 import pytest
 
 from merino.middleware.geolocation import Location
 from merino.providers.base import SuggestionRequest
-
-SuggestionRequestFixture = Callable[[str], SuggestionRequest]
+from tests.unit.types import SuggestionRequestFixture
 
 
 @pytest.fixture(scope="session", name="srequest")
